requirements.txt
Django==5.2.7
Pillow==12.0.0
django-phonenumber-field==8.4.0
phonenumbers==9.0.23
